% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRGibbs.R
\name{ComputeProb}
\alias{ComputeProb}
\title{Generate LogLike of r}
\usage{
ComputeProb(theta_array, matrix_z, idx_m, single_label, X, y)
}
\arguments{
\item{theta_array}{the defined theta above}

\item{matrix_z}{the z matrix}

\item{idx_m}{the mth model of the subpopulation}

\item{single_label}{the chosen variable to be decided whether to be included in the mth model}

\item{X}{the independent variables}

\item{y}{the dependent variable}
}
\value{
updated theta_array
}
\description{
Generate LogLike of r
}
\examples{
\dontrun{
X<-matrix(rep(0,100),nrow=10,ncol=10)
y<-rep(1,100)
initialized<-init(3,X,y)
theta_array<-initialized$theta
new_Z<-UpdateArrayZ(theta_array,y,X,3)
prob<-ComputeProb(theta_array,new_Z, 1,1,X)
}
}
