% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRGibbs.R
\name{GetPosteriorZ}
\alias{GetPosteriorZ}
\title{Update Z:only one scalar at a time}
\usage{
GetPosteriorZ(ym, Xm, theta_array, M)
}
\arguments{
\item{ym}{the dependent variables}

\item{Xm}{the independent variables}

\item{theta_array}{the defined theta above}

\item{M}{the number of subpopulation}
}
\value{
an scalar of z
}
\description{
Update Z:only one scalar at a time
}
\examples{
\dontrun{
X<-matrix(rep(0,100),nrow=10,ncol=10)
y<-rep(1,100)
initialized<-init(3,X,y)
theta_array<-initialized$theta
GZ<-GetPosteriorZ(y[1],X[1,],theta_array,3)
}
}
