% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRGibbs.R
\name{init}
\alias{init}
\title{Initialize all the parameters}
\usage{
init(X, y, M)
}
\arguments{
\item{X}{the independent variables}

\item{y}{the dependent variables}

\item{M}{the number of subpopulation}
}
\value{
an array of theta and z
}
\description{
Initialize all the parameters
}
\examples{
\dontrun{
X<-matrix(rep(0,100),nrow=10,ncol=10)
y<-rep(1,100)
initialized<-init(3,X,y)
}
}
