% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRGibbs.R
\name{init_r}
\alias{init_r}
\title{Initialize r}
\usage{
init_r(X, M)
}
\arguments{
\item{X}{the independent variables}

\item{M}{the number of subpopulation}
}
\value{
a matrix
}
\description{
Initialize r
}
\examples{
\dontrun{
X<-matrix(rep(0,100),nrow=10,ncol=10)
vector_r_0 <- init_r(X,3)
}
}
