% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRGibbs.R
\name{init_rho}
\alias{init_rho}
\title{Initialize rho}
\usage{
init_rho(M)
}
\arguments{
\item{M}{the number of subpopulations}
}
\value{
a vector of size
}
\description{
Initialize rho
}
\examples{
\dontrun{
vector_rho_0 <- init_rho(3)
}
}
