% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRGibbs.R
\name{init_theta}
\alias{init_theta}
\title{Initialize theta}
\usage{
init_theta(M, matrix_z_0, matrix_r_0, vector_rho_0, X, y)
}
\arguments{
\item{M}{the number of subpopulation}

\item{matrix_z_0}{the z matrix}

\item{matrix_r_0}{the r matrix}

\item{vector_rho_0}{the rho vector}

\item{X}{the independent variables}

\item{y}{the dependent variable}
}
\value{
return an array of beta betahat rho sigma oldbeta r
}
\description{
Initialize theta
}
\examples{
\dontrun{
X<-matrix(rep(0,100),nrow=10,ncol=10)
rho <- init_rho(3)
r<-init_r(X,3)
z<-ini_z(X,rho)
theta<-init_theta(3,z,r,rho,X)
}
}
