% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRGibbs.R
\name{init_z}
\alias{init_z}
\title{Initialize z}
\usage{
init_z(X, vector_rho_0)
}
\arguments{
\item{X}{the independent variables}

\item{vector_rho_0}{the rho vector}
}
\value{
a matrix
}
\description{
Initialize z
}
\examples{
\dontrun{
X<-matrix(rep(0,100),nrow=10,ncol=10)
rho<-init_rho(3)
z<-ini_z(X,rho)
}
}
