% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MyEM}
\alias{MyEM}
\title{EM algorithm for the Genetic problem}
\usage{
MyEM(nc, ni, nt)
}
\arguments{
\item{nc}{the initial number of c type}

\item{ni}{the initial value of i type}

\item{nt}{the initial of t type}
}
\value{
the estimated probability
}
\description{
EM algorithm for the Genetic problem
}
\examples{
\dontrun{
p=MyEM(85,96,25)
}
}
