% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRGibbs.R
\name{UpdateArrayZ}
\alias{UpdateArrayZ}
\title{Update the whole vector}
\usage{
UpdateArrayZ(theta_array, y, X, M)
}
\arguments{
\item{theta_array}{the defined theta above}

\item{y}{the dependent variables}

\item{X}{the independent variables}

\item{M}{the number of subpopulation}
}
\value{
an matrix of z
}
\description{
Update the whole vector
}
\examples{
\dontrun{
X<-matrix(rep(0,100),nrow=10,ncol=10)
y<-rep(1,100)
initialized<-init(3,X,y)
theta_array<-initialized$theta
new_Z<-UpdateArrayZ(theta_array,y,X,3)
}
}
