% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRGibbs.R
\name{UpdateR}
\alias{UpdateR}
\title{update r}
\usage{
UpdateR(theta_array, matrix_z, X)
}
\arguments{
\item{theta_array}{the defined theta above}

\item{matrix_z}{the z matrix}

\item{X}{the independent variable}
}
\value{
updated theta_array
}
\description{
update r
}
\examples{
\dontrun{
X<-matrix(rep(0,100),nrow=10,ncol=10)
y<-rep(1,100)
initialized<-init(3,X,y)
theta_array<-initialized$theta
new_Z<-UpdateArrayZ(theta_array,y,X,3)
prob<-UpdateR(theta_array,new_Z)
}
}
