% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRGibbs.R
\name{UpdateSubPop}
\alias{UpdateSubPop}
\title{Update the subpopulation}
\usage{
UpdateSubPop(theta_array, matrix_z, X, M)
}
\arguments{
\item{theta_array}{the defined theta above}

\item{matrix_z}{the z matrix}

\item{X}{the independent variables}

\item{M}{the number of subpopulation}
}
\value{
updated theta_array
}
\description{
Update the subpopulation
}
\examples{
\dontrun{
X<-matrix(rep(0,100),nrow=10,ncol=10)
y<-rep(1,100)
initialized<-init(3,X,y)
theta_array<-initialized$theta
new_Z<-UpdateArrayZ(theta_array,y,X,3)
covariates<-UpdateSubPop(theta_array,new_Z,X,3)
}
}
