% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRGibbs.R
\name{UpdateW_inv}
\alias{UpdateW_inv}
\title{Update the w_inv}
\usage{
UpdateW_inv(theta_array, covariates, M)
}
\arguments{
\item{theta_array}{the defined theta above}

\item{covariates}{the new x subpopulation}

\item{M}{the number of subpopulation}
}
\value{
updated theta_array
}
\description{
Update the w_inv
}
\examples{
\dontrun{
X<-matrix(rep(0,100),nrow=10,ncol=10)
y<-rep(1,100)
initialized<-init(3,X,y)
theta_array<-initialized$theta
new_Z<-UpdateArrayZ(theta_array,y,X,3)
covariates<-UpdateSubPop(theta_array,new_Z,X,3)
theta_array<-UpdateW_inv(theta_array, covariates, 3)
}
}
